♻️ move resourceUtils to resource
