+++ packages/rum/src/typesV2.ts
@@ -1,5 +1,5 @@
 import { Context, ErrorSource, HttpContext, ResourceType } from '@datadog/browser-core'
-import { PerformanceResourceDetailsElement } from './domain/rumEventsCollection/resourceUtils'
+import { PerformanceResourceDetailsElement } from './domain/rumEventsCollection/resource/resourceUtils'
 import { ActionType } from './domain/rumEventsCollection/userActionCollection'
 import { ViewLoadingType } from './domain/rumEventsCollection/viewCollection'
